REFACTOR: #5.4 requestSecret Mutation and Refactor
Added requestSecret mutaion and refactoring auth page
