+++ origin/src/Routes/Auth/AuthQueries.js
@@ -1,0 +1,7 @@
+import { gql } from "apollo-boost";
+
+export const LOG_IN = gql`
+  mutation requestSecret($email: String!) {
+    requestSecret(email: $email)
+  }
+`;
