+++ origin/src/Routes/Auth/index.js
@@ -1,0 +1,3 @@
+import AuthContainer from "./AuthContainer";
+
+export default AuthContainer;
